+++ src/fetchez/streams/__init__.py
@@ -10,7 +10,3 @@ fetchez.streams
 :copyright: (c) 2010-2026 Regents of the University of Colorado
 :license: MIT, see LICENSE for more details.
 """
-
-from .base import BaseReader
-
-__all__ = ["BaseReader"]