+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `cache` command cli to clear cache info
 - Add `--refresh` option to refresh cache at runtime
 - Add `--ignore-failures` option to ignore module and hook exceptions
+- Add `Pipeline` and `read` functions to the api
+- Add `BaseStream` class to fetchez.streams to yield chunks from `Readers`
 
 ### CHANGED
 - Schemas are now for pure validation, mutations happen with modifiers instead
@@ -23,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update modules.base cached_run to use relative paths in cache instead of absolute paths
 - Upgrade Exception handling for 'fail-fast' defaults
 - Update most code to use pathlib, ongoing...
+- Changed `--ignore-failures` to `--fail-fast` (may change back)
+- Moved module/hook expansion from Recipe to registry
 
 ### BUGFIX
  - keyboard interrupt wouldn't correctly get passed to the concurrent threads and got looped into the logging library; moving the try/except to inside the with concurrent and adding a logging.debug in the keyboardinterrupt solves this and now is responsive and acting as expected.